extract ga generation
