+++ tool/cli.ts
@@ -29,7 +29,6 @@ import {
   BUILD_OUT_ROOT,
   CONTENT_ROOT,
   CONTENT_TRANSLATED_ROOT,
-  GOOGLE_ANALYTICS_MEASUREMENT_ID,
 } from "../libs/env/index.js";
 import { runMakePopularitiesFile } from "./popularities.js";
 import { runOptimizeClientBuild } from "./optimize-client-build.js";
@@ -177,14 +176,6 @@ interface PopularitiesActionParameters extends ActionParameters {
     refresh: boolean;
   };
   logger: Logger;
-}
-
-interface GoogleAnalyticsCodeActionParameters extends ActionParameters {
-  options: {
-    measurementId: string;
-    debug: boolean;
-    outfile: string;
-  };
 }
 
 interface BuildRobotsTxtActionParameters extends ActionParameters {
@@ -871,66 +862,6 @@ program
         )
       );
     })
-  )
-
-  .command(
-    "google-analytics-code",
-    "Generate a .js file that can be used in SSR rendering"
-  )
-  .option("--outfile <path>", "name of the generated script file", {
-    default: path.join(BUILD_OUT_ROOT, "static", "js", "gtag.js"),
-  })
-  .option(
-    "--measurement-id <id>[,<id>]",
-    "Google Analytics measurement IDs (defaults to value of $GOOGLE_ANALYTICS_MEASUREMENT_ID)",
-    {
-      default: GOOGLE_ANALYTICS_MEASUREMENT_ID,
-    }
-  )
-  .action(
-    tryOrExit(
-      async ({ options, logger }: GoogleAnalyticsCodeActionParameters) => {
-        const { outfile, measurementId } = options;
-        const measurementIds = measurementId.split(",").filter(Boolean);
-        if (measurementIds.length) {
-          const dntHelperCode = fs
-            .readFileSync(
-              new URL("mozilla.dnthelper.min.js", import.meta.url),
-              "utf-8"
-            )
-            .trim();
-
-          const firstMeasurementId = measurementIds.at(0);
-          const gaScriptURL = `https://www.googletagmanager.com/gtag/js?id=${encodeURIComponent(firstMeasurementId)}`;
-
-          const code = `
-// Mozilla DNT Helper
-${dntHelperCode}
-// Load GA unless DNT is enabled.
-if (Mozilla && !Mozilla.dntEnabled()) {
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-  ${measurementIds
-    .map((id) => `gtag('config', '${id}', { 'anonymize_ip': true });`)
-    .join("\n  ")}
-
-  var gaScript = document.createElement('script');
-  gaScript.async = true;
-  gaScript.src = '${gaScriptURL}';
-  document.head.appendChild(gaScript);
-}`.trim();
-          fs.writeFileSync(outfile, `${code}\n`, "utf-8");
-          logger.info(
-            chalk.green(
-              `Generated ${outfile} for SSR rendering using ${measurementId}.`
-            )
-          );
-        } else {
-          logger.info(chalk.yellow("No Google Analytics code file generated"));
-        }
-      }
-    )
   )
 
   .command(
